+++ demo/api/src/common/blocks/linkBlock/link.block.ts
@@ -1,5 +1,5 @@
-import { EmailLinkBlock, ExternalLinkBlock, PhoneLinkBlock } from "@comet/blocks-api";
-import { createLinkBlock, DamFileDownloadLinkBlock, InternalLinkBlock } from "@comet/cms-api";
+import { ExternalLinkBlock } from "@comet/blocks-api";
+import { createLinkBlock, DamFileDownloadLinkBlock, EmailLinkBlock, InternalLinkBlock, PhoneLinkBlock } from "@comet/cms-api";
 import { NewsLinkBlock } from "@src/news/blocks/news-link.block";
 
 export const LinkBlock = createLinkBlock({
